+++ src/components/resource/components/edit/resource-settings/edit-resource-settings.js
@@ -7,6 +7,8 @@ import hasIn from 'lodash/hasIn';
 import {
   Accordion,
   Icon,
+  Row,
+  Col,
 } from '@folio/stripes/components';
 
 import VisibilityField from '../../../_fields/visibility';
@@ -65,14 +67,28 @@ const EditResourceSettings = ({
     >
       {renderFields
         ? (
-          <>
-            <VisibilityField disabled={visibilityMessage} />
-            <div>
-              {hasInheritedProxy && getAccordionContent()}
-              {model.package.isCustom && <CustomUrlFields />}
-              <AccessTypeEditSection accessStatusTypes={accessStatusTypes} />
-            </div>
-          </>
+          <Row>
+            <Col xs={3}>
+              <VisibilityField disabled={visibilityMessage} />
+            </Col>
+            <Col xs={9}>
+              <Row>
+                {hasInheritedProxy && (
+                  <Col xs={4}>
+                    {getAccordionContent()}
+                  </Col>
+                )}
+                {model.package.isCustom && (
+                  <Col xs={4}>
+                    <CustomUrlFields />
+                  </Col>
+                )}
+                <Col xs={4}>
+                  <AccessTypeEditSection accessStatusTypes={accessStatusTypes} />
+                </Col>
+              </Row>
+            </Col>
+          </Row>
         )
         : (
           <p data-test-eholdings-resource-edit-settings-message>